fix bugs while training with bf16

+++ videolisa/model/sam2/modeling/position_encoding.py
@@ -56,7 +56,7 @@ class PositionEmbeddingSine(nn.Module):
         x_embed = x * self.scale
         y_embed = y * self.scale
 
-        dim_t = torch.arange(self.num_pos_feats, dtype=torch.float32, device=x.device)
+        dim_t = torch.arange(self.num_pos_feats, device=x.device)
         dim_t = self.temperature ** (2 * (dim_t // 2) / self.num_pos_feats)
 
         pos_x = x_embed[:, None] / dim_t
@@ -93,12 +93,12 @@ class PositionEmbeddingSine(nn.Module):
             return self.cache[cache_key].to(device)[None].repeat(B, 1, 1, 1)
 
         y_embed = (
-            torch.arange(1, H + 1, dtype=torch.float32, device=device)
+            torch.arange(1, H + 1, device=device)
             .view(1, -1, 1)
             .repeat(B, 1, W)
         )
         x_embed = (
-            torch.arange(1, W + 1, dtype=torch.float32, device=device)
+            torch.arange(1, W + 1, device=device)
             .view(1, 1, -1)
             .repeat(B, H, 1)
         )
@@ -108,7 +108,7 @@ class PositionEmbeddingSine(nn.Module):
             y_embed = y_embed / (y_embed[:, -1:, :] + eps) * self.scale
             x_embed = x_embed / (x_embed[:, :, -1:] + eps) * self.scale
 
-        dim_t = torch.arange(self.num_pos_feats, dtype=torch.float32, device=device)
+        dim_t = torch.arange(self.num_pos_feats, device=device)
         dim_t = self.temperature ** (2 * (dim_t // 2) / self.num_pos_feats)
 
         pos_x = x_embed[:, :, :, None] / dim_t
@@ -157,7 +157,7 @@ class PositionEmbeddingRandom(nn.Module):
         """Generate positional encoding for a grid of the specified size."""
         h, w = size
         device: Any = self.positional_encoding_gaussian_matrix.device
-        grid = torch.ones((h, w), device=device, dtype=torch.float32)
+        grid = torch.ones((h, w), device=device, dtype=self.positional_encoding_gaussian_matrix.dtype)
         y_embed = grid.cumsum(dim=0) - 0.5
         x_embed = grid.cumsum(dim=1) - 0.5
         y_embed = y_embed / h
@@ -173,7 +173,7 @@ class PositionEmbeddingRandom(nn.Module):
         coords = coords_input.clone()
         coords[:, :, 0] = coords[:, :, 0] / image_size[1]
         coords[:, :, 1] = coords[:, :, 1] / image_size[0]
-        return self._pe_encoding(coords.to(torch.float))  # B x N x C
+        return self._pe_encoding(coords.to(self.positional_encoding_gaussian_matrix.dtype))  # B x N x C
 
 
 # Rotary Positional Encoding, adapted from:
